offbeat_back/backend-master/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-extensions==2.2.9
django-filter==2.3.0
django-hashers-passlib==0.4
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
feedparser==5.2.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
Markdown==3.2.2
newscatcher==0.2.0
oauthlib==3.1.0
passlib==1.7.2
Pillow==7.1.2
pipreqs==0.4.10
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
RestAuthCommon==0.7.0
six==1.15.0
South==1.0.2
sqlparse==0.3.1
tldextract==2.2.2
urllib3==1.25.9
yarg==0.1.9
zipp==3.1.0
django-heroku==0.3.1
django-taggit==1.3.0
django-cleanup==5.1.0
gspread==3.6.0
oauth2client==4.1.3